ONC-1109 ONC-1115 ONC-1122 ONC-1123 ONC-1124 ONC-1125 ONC-1150
git-svn-id: https://source.sakaiproject.org/svn/gradebook/branches/non-cal@61751 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ app/business/src/java/org/sakaiproject/tool/gradebook/business/GradebookManager.java
@@ -436,7 +436,7 @@ public interface GradebookManager {
     * @return id of the new category
     * @throws ConflictingAssignmentNameException StaleObjectModificationException
     */
-    public Long createCategory(final Long gradebookId, final String name, final Double weight, final int drop_lowest, final Integer dropLowest, final Integer dropHighest, final Integer keepHighest, final Double pointValue, final Integer relativeWeight, final Boolean is_extra_credit) 
+    public Long createCategory(final Long gradebookId, final String name, final Double weight, final int drop_lowest, final Integer dropLowest, final Integer dropHighest, final Integer keepHighest, final Double pointValue, final Double relativeWeight, final Boolean is_extra_credit) 
     throws ConflictingCategoryNameException, StaleObjectModificationException;
     
     /**method to get all categories for a gradebook
